Commit adds Digital Media Page / reworks annotation modal code structure
This commit adds a view for the Digital Media Item that can be accessed through it's own unique URL. This page allows the user to see the digital media item and also to click through the other media items associated with the target specimen. Annotating media metadata is possible through the same annotation modal as on the specimen page, additionally a beginning for visual annotating of images has been added.

The annotation modal's code has been reworked a bit so it is easier to include this component in other pages as well. In the future it is probably more usefull to create a seperate folder for these kind of components instead of including it from the annotate folder that is nested in the specimen folder.

Adds:
- Digital Media Item view at /dm/handle
- Viewing chosen image and other target specimen related images
 - Annotating digital media items metadata with the annotation modal
- See annotations through the annotations overview on this page
- Beginning for visual annotations of images and possibly other media items in the future
- Filter layer for digital media object

Modifies:
- Annotation modal, to be more suited for implementation in other components than the specimen page