Review commit: implement Keycloak Mock
This commit is based upon the code review for the PR of this branch.

This commit resolves the Keycloak mock problem within the test suites. The current solution is to use a static string as the token, as well as the Keycloak subject. These can be accessed by the application as normal as the KeycloakService module is mocked by a mock KeycloakService which returns the static values.

As the front-end does not check the validity of the JWA-token and the tests are devoted to the UI with mocked API requests, it is not necessary to create a valid JWA token.