Restructure Searching / Filtering
The main purpose of this commit is to add search / filter capability to the system. Users will now be able to search by PID, Physical ID or Organisation from the Homepage which will redirect. The Search page now also includes the new method for filtering. To set an example, the free text search (q) and organisation (organisationId) are included as possible filters. The filters will need to be generalized when other filter properties are added.

Commit also contains some other stuff that has been worked on in the meantime.

Adds:
- PID, Physical Specimen ID and Organisation Search methods to the Home Page
- Functional Text search and Organisation filters to the Search Page
- Toggle button in Specimen Page to switch to the JSON variant

Modifies:
- Searching will now primarily be done via the filter functionality which makes use of the search params in the url. These params take priority over the Redux state, which used to save the search query. Future changes will need to be made upon this modification
- The Specimen page now uses a generic API call to call on the specimen specific data, digital media and annotations

Bug Fixes:
- Transparent Version bar on Specimen page. Now makes use of new library for enhanced select fields
- Original data and Annotations tables on Specimen page are now scrollable