Expand ID Card feature to Search Page
Commit expands the ID Card feature to the Search Page. It is however, not the same as the current one on the Specimen Page, but based on the same principle. The current search flow: the user can search for specimens by filters, results will show up, the user can select a specimen from the table, the table shrinks and covers the filters, revealing the ID Card containing the primary Specimen details and digital media if so. A button called: 'See more details' leads the user to the full Specimen page. The filters can be shown again by pressing the return button.

Adds:
- ID Card containing primary Specimen information on the Search Page
- Styling based upon designs to results table and ID card on the Search Page
- Functionality to switch between filters and the ID Card by shrinking the columns in size and or positioning them differently with the help of position absolute