Base Paginator on 'links' data from API
Commit bases Paginator component based on the returned 'links' data from the API containing information about the next, current and previous pages. This information is more reliable and easy going than the method committed in the previous commit. It is also less intensive since the API is now only called once and not twice.

Modifies:
- Paginator to make use of the 'links' data from the API to determine page nubers

Removed:
- The pagination range variable, the Paginator now solely works based on the page number