Add Coverage Report and resolve Testing with modules
Commit adds the coverage report of Jest (testing library) to be used in the Sonar Cloud report, this edits the GitHub actions to include the installation of dependencies to run the test cases with coverage. Testing was broken due to the inclusion of libraries like 'react-leaflet' that use a module like approach. Resolved this by adding a bit of Jest configuration to the package.json.

Adds:
- Coverage Report for use in the Sonar Cloud analysis / reportage

Modifies:
- Mock data for tests to make use of the JSON API standard
- Mock API to be mocked in global AppRenderer file instead of each separate test file

Resolves:
- Testing bug that causes tests over certain files with module library inclusions to fail by adding some Jest configuration in package.json