Rework filters and add filter tags above results table
Commit reworks the filters on the Search Page according to the design and improves functionality; as well as diversity in filters. Multi select filters will still have the checkboxes, an exception in styling is made for the MIDS filter. Date filters use date pickers, but are not yet integrated into the search functionality. For free text fields it is the same thing.

The filter tags above the results table indicate which filters have been chosen, also when the filters are collapsed. Pressing the x icon of a tag will remove the specific filter from the filters list, without having to search for it in the filters list.

Adds:
- Improved Multi Select Filter and option
- Style wise a different option for the MIDS multi select filter
- Date filter field with datepicker
- Free text filter field
- Filter tags above the results table for display when filters sidebar is closed

Modifies:
- Reworks the styling of the filters sidebar

Resolves:
- Multi select of filters not working appropriate since the search params were messed up by always removing the last value of a filter, or removing all of the values from a filter