Add active filters (dropdown) list and fix index issue with select boxes
Commit adds the active filters as tags above the results table, they are always visible, pressing the x icon on the tag will remove the filter from the active filters list and search params. Also resolves the issue of indexing select boxes wrong so index 1 could delete index 0 if the order of aggregations allowed this.

Adds:
- Active filters tags
- Active filters dropdown

Resolves:
- Index issue of select option checkboxes