Fix back button issue on Search Page
Commit fixes the back button issue on the Search. The issue caused for every filter on a new render to add a (empty) route to the history which stacked up a lot of unnecessary routes and back clicks. Now, it only pushes for the ones that actually trigger a change in the length of the chosen filters and thus search params. Hence, hitting the back button will first remove or readd filter based on the history, or return back to the page before (often Home)