Add basic Annotations Side Panel
Commit adds the basic Annotations Side Panel. The panel can be opened for all annotations of a target, or for a specific field of the target. Annotations are ordered by motivation type. Using the motivation filter, you can filter on the different types. The Side Panel makes the default page shrink, which is slightly compensated by enlarging the width of the default page, making it closer to the edges of the screen (like in the design). This also affects the header and footer components.

Adds:
- Basic annotations side panel
- Ordered annotations by motivation type
- Filter on motivation type using the motivation type select filter

Modifies:
- Disables the old annotation modal, will be removed in the future