Patch and Delete Annotations
Commit readds the functionality to patch and delete annotations. A generic Update function will make sure the side panel is updated everytime. A separate edit annotation variable is defined to store the annotation being edited. This variable is emptied when closing the form, side panel or refreshing the page.

Adds:
- Patch annotation functionality
- Delete annotation functionality
- Generic update function for the annotations overview (component)