Add refresh button to Side Panel in order to refresh Annotations
Commit adds a refresh button to the Side Panel that can be used to refresh the annotations within it. This method is dependent upon the parent component passing it on, otherwise the button will not be displayed. The Specimen of Digital Media component will trigger a refresh call to the API and provide the annotations to the side panel, as well as update the source material.

Adds:
- Refresh functionality to the Side Panel
- Refresh functions for Specimen and Digital Media Annotations

Modifies:
- ShowWithAllAnnotations function to become a ShowWithAnnotations function that takes two parameters: optional annotations (default is all annotations), and a target property to keep track of).