Resolve Version switch Bug
Commit resolves version switch Bug.
The Bug was that the array slicing of the path variable, to set the version on the url of a specimen, after selecting a specific version, only looked at the last two indexes for the dash (/)  and a version number. However, all versions above ten have two digits, making this method incorrect.

Resolved this by using the lastIndexOf method which detects the index of the dash and then slices the url untill the end, no matter the amount of digits.