Modify steps process and add steps for Search Page
Commit modifies the process of handling the intro.js steps. Now loaded via an external component, makes these steps less prone to modifications in the original file, and prevents a lot of lines of code in the original file.

Commit also adds steps for the Search page which give a demo of the Search and filter functionality.

Adds:
- Search Page intro steps
- Possibility for multiple sets of steps on the same page

Modifies:
- Steps handling process