Remember pagination number
Commit adds the following behavior to the Paginator: the paginator keeps track of a new pagination object in Redux state that holds: the page (home, search, specimen, etc.), pageNumber of search and possible filters. Onchange of the search results and pagination pages, the object will be updated by the paginator. This will directly affect the behavior of the Search page to for example stick to the same pagination page when navigating between different pages (when not refreshing). This should make it easier to keep track of what you were searching on.

The new pagination object could be used for multiple purposes in the future, as it also holds the filters in state.

Adds:
- Pagination object to Redux state

Modifies:
- Searching behavior of Search page using the paginator and in combination with filters