Update geology topic discipline, take a tour button, download file name with underscore and change annotation motivation to type
Commit changes several small things for improving the user experience. It changes the Earth System topic discipline to geology. Adds a 'take a tour' button to the header to initiate a tour rather than the compass icon. Changes the downloaded specimen file name to be the PID with the version attached. Lastly, changes annotation motivation to annotation type because the term motivation would be too vague.

Adds:
- Take a tour button instead of the compass icon to start a tour on one of the available pages

Modifies:
- Earth System becomes Geology
- Changes the downloaded Specimen file name to contain the PID and version after an underscore
- The term annotation motivation to annotation type