Fix Github issue #55
Commit should fix Github issue #55. Homepage is a unique scenario in which the page should scroll if the content gets too big for the screen. The Footer gets pushed down to the ultimate bottom and is not always visible unlike the header.

This behavior is not applied to other pages.