Add Image Annotation
Commit adds Image Annotation to the Digital Media page.

The user will be able, when logged in, to make visual annotations upon a 2D image using the Annotorious tool. The tool is embedded into the OpenSeaDragon viewer, nog also making it possible to zoom in and pan a 2D image to full screen mode, just like the IIIF images.

You can add, edit and delete visual annotations. At the moment, there is no user check, so everyone can edit each others annotations. Annotations can be made by holding the shift button, clicking and dragging the mouse across the image, or by clicking on the make an annotation button, clicking on the image and dragging the mouse across the image.

The visual images are also present within the side panel. When making, editing or deleting a visual annotation, with the side panel open, will not directly update the side panel yet.

Adds:
- Visual image annotation
- Add, edit and delete visual image annotations
- OpenSeaDragon viewer for 2D images