Implement new Data Model Types
Commit implements the new Data Model Types. These types are now generated via a new package, basing the data structure upon JSON Schema Files. The function generates the type files, which are then used within the application, or combined with custom types to broaden them. For example. the DigitalSpecimen and DigitalMedia types are enhanced custom types based upon the base DigitalMedia and DigitalEntity types from the data model.

After the type and structure change of all the source data. A lot of files had to be reviewed to resolve errors within property calls or assignments. Handle links for Specimen and Digital Media have been set to DOI, so the https prefix of DOI is now removed. Basically, a lot of lines had to be slightly changed to adapt to the new data model in a way DiSSCover (kinda) works and displays no errors on deployment.

New Pull Requests to come will take care of all the new incoming fields within the improved data model.

Adds:
- Auto generated type files covering the new data model, based upon the JSON schemas of the model

Modifies:
- Loads of files to accept the new typings and properties of the new data model