Scale Specimen ID Card and Text / change the Digital Specimen and Media types
Commit scales the Specimen ID Card to fit its properties, the height is divided equally. Fixed the banner image not showing up, and also implemented a scalar for the general text in the application to make the interface appear the same, as well on smaller as bigger screens.

Commit also changes the digital specimen and media types. They now include the different levels of nesting as present in the API. This does require  an extra level of assigning or requesting properties, but should be better as it is inline with the source, namely: the API