Add Entity Relationships Tab
Commit adds Entity Relationships tab containing two different views: a table view for displaying all the data in an orderly manner, and a graph view for visually indicating the relations to the specimen. A relation can be selected for more information. This still needs to be filled. The styling also needs to be reviewed.

Adds:
- Entity Relationships tab
- Table view for entity relationships
- Graph view for entity relationships