Add global definition for doi.org and handle.net
Commit adds global definitions in the .env file for doi.org and handle.net as these strings are repeated a lot in the overall code. Now, the variable takes care of it. Also, if in any case the id's will not be DOIs and hosted somewhere else, the url can be changed rapidly.

Adds:
- Global definitions for doi.org and handle.net variables