Resolve Annotation functionality for Demo
Resolve biggest Annotation functionality for demoing purposes. There is still a bunch that needs to be done to complete the whole process as certain features of annotations are invalid, they do not break anything, but are not valid annotation objects. Also the UI needs to be updated for specific annotation cases.