Fix next page bug while searching with filters and improve search logic
Fixes the bug that prevented the user from going to the next page in the search results table.
Improves the search logic so when an user goes to a specimen page from the results table, and immediately returns to the search page, the search results don't get refreshed so that the selected specimen stays in place in the search results table, aggregations and paginator do refresh.