Include Phylopic/ topic discipline icons in search results
Commit includes Phylopic or topic discipline icons in the search results table. Either the phylopic icon can be succesfuly fetched, or the topic discipline one is used. For the Phylopic query, the order, family, genus or scientific name is used, the first one that contains a value, in that order.

Adds:
- Phylopic or topic discipline icons to search results