Add Natural Origin as block to Home
Commit adds the Natural Origin as a button to the Home page, removing the static title. This button will now trigger all the topic disciplines instead of the select all checkbox. Natural origin covers all topic disciplines in count and excludes the unclassified, as this is another topic origin

Modifies:
- Natural Origin to a button on Home that selects all topic disciplines

Removes:
- Select all checkbox on Home