Add Tan Stack table / Create a custom Data Table component with dynamic widths and column pinning
Commit adds Tan Stack table as the new table dependency to replace the react table component dependency as it lacked features. This new dependency is headless, meaning it allows for a lot of custom behavior. This costs more time to implement, but grants a lot of freedom with a light weight approach.

Started a custom component for the data tables within DiSSCover. This component now includes dynamic widths of the columns based upon the viewport and the ability to pin a certain amount of columns to the left of the table.
Also this custom table now has sticky headers which are always visible when scrolling.

Adds:
- TanStack table library
- Custom Datatable component
- Dynamic columns widths to datatable component
- Pinned columns to datatable component