Improve Annotorious set-up
Commit improves the Annotorious setup by a lot. All functions have been revisited and reformatted to make use of the internal Annotorious API instead of external state variables. I believe this is what caused the constant errors whilst trying to annotate. Annotorious losing the trace of annotations because they were not saved internally, but externally.
Also fixed most of the typing issues for it, one remains, forwarded the issue to the Annotorious developer.

Modifies:
- Annotorious related code to make more use of the interal API and remove external variables for tracing the annotations upon the canvas