Reimplement whole annotation and fix edit annotation including current values
Commit reimplements the whole annotation, now called an annotation upon a 'super class'. A super class would be the Digital Specimen or Digital Media. It affects the rendering of certain elements in the view, this can be optimized in the future to make it cleaner (it has become a bit messy).

Commit also fixes the issue when editing a class annotation not showing the individual class fields. The fields are now also filled with the current values upon editing, and when adding a new annotation upon a field or class, the values will be taken over from the current ones that are in the object itself.

Also the back action has be tweaked as when you return from the annotation form, while a field or class is selected, only annotations belonging to that field or class are shown.