Toggle 'Make Annotation' button based upon the sate of the OSD canvas
Commit adds that if the image could not be loaded correctly, due to CORS or something else, the user will not be allowed to make visual annotations use the 'Make Annotation' button. This is done by attaching an event handler to the open method of the OpenSeaDragon Viewer which calls allows the button to spawn if the Viewer succesfully displays the image.