Promt message on completing MAS
Commit adds a check in the MAS polling part for if a scheduled MAS in memory was completed, if so, show a prompt message saying its done.

This behavior is not going to work when the table holds more than 25 records and the record scheduled is not present in the returned array. To make this fully comprehensible, we need to implement Web Sockets.