Add InputListSelect component and use this to facilitate the Taxonomic Species filter
Adds a new component called the Input List Select, which allows for displaying a list of items, completely dependent on the parent's data. The Taxonomic filter uses this component so users can search for specific species by providing a query, the results of this are displayed as the items in the Input List Select. Items can be selected  using the mouse or keyboard buttons, selecting one will activate the species filter for the chosen list option.

Adds:
- Input List Select Component
- Species search and select list to taxonomic filter