Implement annotation wizard functionalities
Implements the required functionalities for the annotation wizard like tracking the current selected index and how far a user has gotten. Steps should be reachable from different positions. A user should be able to go back to previous steps or be able to visit later steps when the requirements are met and the step has been visited before. The behavior is there, now the remaining part is to implement the requirements and functionality per step to make the full process work.