Add functionality to add new instances of super classes, classes or terms with a parent check
Adds the functionality to set a new super class, class or term as the annotation instance. This checks for parent classes so no broken chains will be present. The user will be guided in adding missing parent classes as when a parent class misses, it will need to be added first before the child can be added. But for convenience, it will be possible to add the child simultaneously with the parent class, preventing the user from needing to complete multiple annotations (and as non accepted annotations are not part of the specimen yet, they could not even be targeted).