Add basics of third annotation form step with form segment
Adds the basics of the third annotation form step which renders all of the form fields adhering to the selected annotation target. The form segment component is unique in a way that it can render both objects and arrays that are stored within the selected schema. The schema loaded can be nested inside an existing schema file, so a check for that is included so when its the case, it will load the properties from the parent schema file.