Process and Submit Annotations
Commit adds the functionality to process and submit annotations to the API when a user clicks on the 'save' button at the end of the annotation wizard. The processing and creation of the annotation object are handled as utility functions. The form submit function posts the annotation towards the API and catches the result, going back to the annotation overview with the newly added annotation, or displays an error message when an error occurs during the connection with the API. A new exception has been added for failing POST requests.