Reintegrate the mobile interpreter
Reintegrates the mobile interpreter. It is based on the type of browser the user is using, so no screen width. If a comprehensive mobile version is wanted in the future, this needs to be further defined as it needs its own best practices and guide lines for responsiveness and custom classes in app.css.