Fix issue with child objects and terms in sub array classes in annotation wizard
Fixes issue with child objects and terms lacking the correct index of sub arrays in their path. Indexes are now appended to the path before the suffix of the targeted JSON path. It takes the base of the parent, checks for the index and appends it if there and finally appends the suffix to the field name.