Change the boot structure of the application to create a safe net for when Keycloak fails
Change the boot structure of the application to create a safe net for when Keycloak fails.

Boot function is now called in index as part of the main start up process. It includes keycloak initialization. While booting, the loading screen will be displayed if it takes too long. The boot state (aggregations and phylopic build) will be received during this process and be set in the app.tsx file. If keycloak fails, it will still boot the application